hal: openisa: add files/config to enable vega radio config
Add the necessary files to enable the configuration of the
GenFSK LL, as it's present in the Vega RISCV platform.

The memory map for the RV32M1 ri5cy core has been updated
with the XCVR registers and additional defines, as they
were included in the SDK for the zero-riscy core. The
location of file that includes the added defines is here:

   vega_sdk_riscv/devices/RV32M1/RV32M1_zero_riscy.h

The baseline for the updates done in this commit are
based on SDK Release Version 2.2.0.

Signed-off-by: Radu Alexe <radu.alexe@nxp.com>
Signed-off-by: David Leach <david.leach@nxp.com>
Signed-off-by: Ionut Ursescu <ionut.ursescu@nxp.com>